TweetIngest/requirements.txt
tweepy
textblob
azure-eventhub

